refactor: outosrce and move files, react children usage
