+++ apps/next/src/pages/AccountManagement/components/AccountDetails/components/AddressesCard.tsx
@@ -1,4 +1,4 @@
-import { AddressItem } from '@/pages/AccountManagement/components/AddressItem';
+import { AddressItem } from '@/components/Address/AddressItem';
 import {
   BitcoinColorIcon,
   CChainIcon,
