+++ apps/next/src/popup/index.tsx
@@ -12,7 +12,7 @@ import {
   FeatureFlagsContextProvider,
   SettingsContextProvider,
 } from '@core/ui';
-import { lazy, ReactElement, Suspense } from 'react';
+import { Children, lazy, ReactElement, ReactNode, Suspense } from 'react';
 import { I18nextProvider } from 'react-i18next';
 import { HashRouter as Router } from 'react-router-dom';
 import { initI18n, i18next } from '@core/common';
@@ -28,22 +28,15 @@ const App = lazy(() => {
 });
 
 const root = createRoot(document.getElementById('popup') as HTMLElement);
-type ProviderElement = ReactElement<
-  unknown,
-  string | React.JSXElementConstructor<any>
->;
 
 export const Providers = ({
   providers,
   children,
 }: {
-  providers: ProviderElement[];
-  children: ProviderElement;
+  providers: ReactNode[];
+  children: ReactElement;
 }) => {
-  const renderProvider = (
-    renderedProviders: ProviderElement[],
-    renderedChildren: ProviderElement,
-  ) => {
+  const renderProvider = (renderedProviders, renderedChildren) => {
     const [provider, ...restProviders] = renderedProviders;
 
     if (provider) {
@@ -64,17 +57,14 @@ browser.tabs.query({ active: true }).then(() => {
     <Sentry.ErrorBoundary>
       <Router>
         <Providers
-          providers={[
-            <I18nextProvider i18n={i18next} key={0} />,
-            <ConnectionContextProvider
-              LoadingComponent={CircularProgress}
-              key={1}
-            />,
-            <SettingsContextProvider key={2} />,
-            <FeatureFlagsContextProvider key={3} />,
-            <AnalyticsContextProvider key={4} />,
-            <Suspense fallback={<CircularProgress />} key={5} />,
-          ]}
+          providers={Children.toArray([
+            <I18nextProvider i18n={i18next} />,
+            <ConnectionContextProvider LoadingComponent={CircularProgress} />,
+            <SettingsContextProvider />,
+            <FeatureFlagsContextProvider />,
+            <AnalyticsContextProvider />,
+            <Suspense fallback={<CircularProgress />} />,
+          ])}
         >
           <App />
         </Providers>
